Move client service to own package
