+++ server/api_handler_clients.go
@@ -14,6 +14,7 @@ import (
 	"github.com/cloudradar-monitoring/rport/server/auditlog"
 	"github.com/cloudradar-monitoring/rport/server/clients"
 	"github.com/cloudradar-monitoring/rport/server/clients/clienttunnel"
+	"github.com/cloudradar-monitoring/rport/server/clientservice"
 	"github.com/cloudradar-monitoring/rport/server/ports"
 	"github.com/cloudradar-monitoring/rport/server/routes"
 	"github.com/cloudradar-monitoring/rport/server/validation"
@@ -69,7 +70,7 @@ func convertToClientsPayload(clients []*clients.CalculatedClient, fields []query
 func convertToClientPayload(client *clients.CalculatedClient, fields []query.FieldsOption) ClientPayload { //nolint:gocyclo
 	requestedFields := query.RequestedFields(fields, "clients")
 	p := ClientPayload{}
-	for field := range clientsSupportedFields["clients"] {
+	for field := range clientservice.OptionsSupportedFields["clients"] {
 		if len(fields) > 0 && !requestedFields[field] {
 			continue
 		}
@@ -173,7 +174,7 @@ func getCorrespondingSortFunc(sorts []query.SortOption) (sortFunc func(a []*clie
 
 func (al *APIListener) handleGetClient(w http.ResponseWriter, req *http.Request) {
 	options := query.GetRetrieveOptions(req)
-	errs := query.ValidateRetrieveOptions(options, clientsSupportedFields)
+	errs := query.ValidateRetrieveOptions(options, clientservice.OptionsSupportedFields)
 	if errs != nil {
 		al.jsonError(w, errs)
 		return
@@ -261,8 +262,8 @@ func (al *APIListener) handlePostClientACL(w http.ResponseWriter, req *http.Requ
 }
 
 func (al *APIListener) handleGetClients(w http.ResponseWriter, req *http.Request) {
-	options := query.NewOptions(req, nil, nil, clientsListDefaultFields)
-	errs := query.ValidateListOptions(options, clientsSupportedSorts, clientsSupportedFilters, clientsSupportedFields, &query.PaginationConfig{
+	options := query.NewOptions(req, nil, nil, clientservice.OptionsListDefaultFields)
+	errs := query.ValidateListOptions(options, clientservice.OptionsSupportedSorts, clientservice.OptionsSupportedFilters, clientservice.OptionsSupportedFields, &query.PaginationConfig{
 		MaxLimit:     500,
 		DefaultLimit: 50,
 	})
