+++ server/api_handler_commands_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/cloudradar-monitoring/rport/server/cgroups"
 	"github.com/cloudradar-monitoring/rport/server/chconfig"
 	"github.com/cloudradar-monitoring/rport/server/clients"
+	"github.com/cloudradar-monitoring/rport/server/clientservice"
 	"github.com/cloudradar-monitoring/rport/server/test/jb"
 	"github.com/cloudradar-monitoring/rport/share/comm"
 	"github.com/cloudradar-monitoring/rport/share/logger"
@@ -272,7 +273,7 @@ func TestHandlePostCommand(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
 			// given
-			clientService := NewClientService(nil, nil, clients.NewClientRepository(tc.clients, &hour, testLog))
+			clientService := clientservice.New(nil, nil, clients.NewClientRepository(tc.clients, &hour, testLog))
 			al := APIListener{
 				insecureForTests: true,
 				Server: &Server{
@@ -641,7 +642,7 @@ func TestHandlePostMultiClientCommand(t *testing.T) {
 			al := APIListener{
 				insecureForTests: true,
 				Server: &Server{
-					clientService: NewClientService(nil, nil, clients.NewClientRepository([]*clients.Client{c1, c2, c3}, &hour, testLog)),
+					clientService: clientservice.New(nil, nil, clients.NewClientRepository([]*clients.Client{c1, c2, c3}, &hour, testLog)),
 					config: &chconfig.Config{
 						Server: chconfig.ServerConfig{
 							RunRemoteCmdTimeoutSec: defaultTimeout,
@@ -1901,7 +1902,7 @@ func makeAPIListener(
 	clientRepo *clients.ClientRepository,
 	defaultTimeout int,
 	testLog *logger.Logger) (al *APIListener) {
-	clientService := NewClientService(nil, nil, clientRepo)
+	clientService := clientservice.New(nil, nil, clientRepo)
 	al = &APIListener{
 		insecureForTests: true,
 		Server: &Server{
