+++ server/upload.go
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	errors2 "github.com/cloudradar-monitoring/rport/server/api/errors"
+	"github.com/cloudradar-monitoring/rport/server/clientservice"
 	errors3 "github.com/cloudradar-monitoring/rport/share/errors"
 
 	"github.com/cloudradar-monitoring/rport/server/api"
@@ -88,7 +89,7 @@ func (al *APIListener) handleFileUploads(w http.ResponseWriter, req *http.Reques
 		return
 	}
 
-	cr := ClientServiceProvider{}
+	cr := clientservice.Provider{}
 	clientGroups, err := al.clientGroupProvider.GetAll(req.Context())
 	if err != nil {
 		al.jsonError(w, err)
